Add argument check to index properly.
On Windows the shebang/hashbang is not honored, we need some way to determine what order the arguments are in. Instead of grabbing an argument parsing library (still a possibility) we can just check to see if node has been used to run the script. Not ideal, but it is the best option I can see right now.